Update C# and Java
