+++ csharp/src/Ice/Util.cs
@@ -2,8 +2,6 @@
 
 #nullable enable
 
-using System.Globalization;
-
 namespace Ice;
 
 /// <summary>
@@ -303,76 +301,6 @@ public sealed class Util
     /// <returns>The Ice version.</returns>
     public static int intVersion() => 30850; // AABBCC, with AA=major, BB=minor, CC=patch
 
-    /// <summary>
-    /// Converts a string to a protocol version.
-    /// </summary>
-    /// <param name="version">The string to convert.</param>
-    /// <returns>The converted protocol version.</returns>
-    public static ProtocolVersion stringToProtocolVersion(string version)
-    {
-        stringToMajorMinor(version, out byte major, out byte minor);
-        return new ProtocolVersion(major, minor);
-    }
-
-    /// <summary>
-    /// Converts a string to an encoding version.
-    /// </summary>
-    /// <param name="version">The string to convert.</param>
-    /// <returns>The converted encoding version.</returns>
-    public static EncodingVersion stringToEncodingVersion(string version)
-    {
-        stringToMajorMinor(version, out byte major, out byte minor);
-        return new EncodingVersion(major, minor);
-    }
-
-    /// <summary>
-    /// Converts a protocol version to a string.
-    /// </summary>
-    /// <param name="v">The protocol version to convert.</param>
-    /// <returns>The converted string.</returns>
-    public static string protocolVersionToString(Ice.ProtocolVersion v) => majorMinorToString(v.major, v.minor);
-
-    /// <summary>
-    /// Converts an encoding version to a string.
-    /// </summary>
-    /// <param name="v">The encoding version to convert.</param>
-    /// <returns>The converted string.</returns>
-    public static string encodingVersionToString(Ice.EncodingVersion v) => majorMinorToString(v.major, v.minor);
-
-    private static void stringToMajorMinor(string str, out byte major, out byte minor)
-    {
-        int pos = str.IndexOf('.', StringComparison.Ordinal);
-        if (pos == -1)
-        {
-            throw new ParseException($"malformed version value in '{str}'");
-        }
-
-        string majStr = str[..pos];
-        string minStr = str[(pos + 1)..];
-        int majVersion;
-        int minVersion;
-        try
-        {
-            majVersion = int.Parse(majStr, CultureInfo.InvariantCulture);
-            minVersion = int.Parse(minStr, CultureInfo.InvariantCulture);
-        }
-        catch (FormatException ex)
-        {
-            throw new ParseException($"invalid version value in '{str}'", ex);
-        }
-
-        if (majVersion < 1 || majVersion > 255 || minVersion < 0 || minVersion > 255)
-        {
-            throw new ParseException($"range error in version '{str}'");
-        }
-
-        major = (byte)majVersion;
-        minor = (byte)minVersion;
-    }
-
-    private static string majorMinorToString(byte major, byte minor) => $"{major}.{minor}";
-
-    public static readonly ProtocolVersion Protocol_1_0 = new(1, 0);
     public static readonly EncodingVersion Encoding_1_0 = new(1, 0);
     public static readonly EncodingVersion Encoding_1_1 = new(1, 1);
